Return local file paths with cached keys (#30)
This change extends the knox.Key structure to include a Path field that
will contain the local path of the cached key file for registered keys.
The field will be empty and omitted when the key is fetched remotely.

This is useful for applications that register keys and then watch the
cached files for changes. Previously, these applications needed to know
both the root path (e.g., /var/lib/knox/v0/keys/) and naming format of
cached files. Now they can simply read the path directly from the
registration response.