Add Principal Mux to hold multiple principals
This way the Authentication middleware can provide the mux of all
valid principals that can be constructed. Then the downstream handler
that checks CanAccess on that set of principals can check each one
and allow access if any of them would grant access.

This is really only possible because of the SpiffeFallbackProvider,
because the only way that a call could have two valid Principals
is if you are using that provider as well as the MachineAuth provider.

I wrestled with where this mux should be defined. Particularly because
of the changes needed in auth.go to support type inspection into the
mux for IsService and IsUser. Ultimately, I concluded that it belonged
next to the Principal interface itself as that is the only thing that
it is actually dependent on.

The overall goal is to provide access to keys by either machine or
service auth so that we can begin granting key priviledges using service
without having to move all keys over whole sale.